EDU-1074 Extract user steps component for registration and reset password flows
