+++ src/components/registration-form.less
@@ -1,28 +1,7 @@
 .RegistrationForm {}
 
-.RegistrationForm-steps {
-  margin-bottom: 50px;
-}
-
-.RegistrationForm-step {
-  font-size: @edu-font-size-sm !important;
-
-  &.is-selected {
-    font-weight: bold;
-  }
-}
-
 .RegistrationForm-termsAndConditionsCheckbox {
   &.ant-checkbox-wrapper {
     line-height: initial;
   }
-}
-
-.RegistrationForm-registerButton {
-  .m-user-form-submit-button;
-}
-
-.RegistrationForm-verificationMessage {
-  text-align: center;
-  margin-bottom: 50px;
 }
