+++ src/components/user-steps-form.less
@@ -1,0 +1,26 @@
+.UserStepsForm {
+  .ant-result {
+    padding: 0 !important;
+  }
+}
+
+.UserStepsForm-steps {
+  margin-bottom: 50px;
+}
+
+.UserStepsForm-step {
+  font-size: @edu-font-size-sm !important;
+
+  &.is-selected {
+    font-weight: bold;
+  }
+}
+
+.UserStepsForm-button {
+  .m-user-form-submit-button;
+}
+
+.UserStepsForm-verificationMessage {
+  text-align: center;
+  margin-bottom: 50px;
+}
