+++ src/components/user-steps-form.yml
@@ -1,0 +1,30 @@
+step_enterData:
+  en: Enter data
+  de: Daten eingeben
+step_enterCode:
+  en: Confirm
+  de: Bestätigen
+step_completed:
+  en: Done
+  de: Fertig
+verificationCode:
+  en: Verification code
+  de: Verifizierungs-Code
+enterVerificationCode:
+  en: Please enter your verification code
+  de: Bitte gib deinen Verifizierungscode ein
+verificationCodeIsInvalid:
+  en: The verification code is invalid or expired
+  de: Der Verifizierungscode ist ungültig oder abgelaufen
+verificationMessage:
+  en: |
+    We have sent you an email with a verification code. Please copy this code into the textbox below and confirm your input.
+
+    The code is valid **for {validityInMinutes}**.
+  de: |
+    Wir haben dir eine E-Mail mit einem Verifizierungs-Code gesendet. Bitte kopiere diesen Code in das Textfeld unten und bestätige deine Eingabe.
+
+    Der Code ist **{validityInMinutes} lang** gültig.
+gotoDashboard:
+  en: Show my personal dashboard
+  de: Mein persönliches Dashboard anzeigen
