+++ src/styles/main.less
@@ -60,6 +60,7 @@
 @import '../components/pages/dashboard.less';
 @import '../components/section-display.less';
 @import '../components/iteration-panel.less';
+@import '../components/user-steps-form.less';
 @import '../components/sections-display.less';
 @import '../components/sorting-selector.less';
 @import '../components/copyright-notice.less';
